+++ .env
@@ -1,2 +1,2 @@
-DB_URL=./src/db/db.sqlite
+DB_URL=db.sqlite
 NEXTAUTH_URL=http://localhost:3000/
