+++ drizzle.config.ts
@@ -1,9 +1,8 @@
 import type { Config } from "drizzle-kit"
-import dotevn from "dotenv"
-dotevn.config()
+import "dotenv/config"
 
 export default {
-  schema: "./src/db/schema.ts",
+  schema: "./src/lib/db/schema/**/*",
   out: "./src/db/migrations",
   driver: "better-sqlite",
   dbCredentials: {
