+++ src/lib/db/index.ts
@@ -1,0 +1,5 @@
+import { drizzle, type BetterSQLite3Database } from 'drizzle-orm/better-sqlite3'
+import Database from 'better-sqlite3'
+
+const sqlite = new Database('sqlite.db')
+export const db: BetterSQLite3Database = drizzle(sqlite)
