+++ src/lib/db/migrate.ts
@@ -1,0 +1,35 @@
+import "dotenv/config"
+
+import { drizzle, type BetterSQLite3Database } from "drizzle-orm/better-sqlite3"
+import { migrate } from "drizzle-orm/better-sqlite3/migrator"
+import Database from 'better-sqlite3'
+
+
+const runMigrate = async () => {
+  if (!process.env.DATABASE_URL) {
+    throw new Error("DATABASE_URL is not defined")
+  }
+
+
+  const sqlite = new Database('sqlite.db')
+  const db: BetterSQLite3Database = drizzle(sqlite)
+
+
+  console.log("⏳ Running migrations...")
+
+  const start = Date.now()
+
+  await migrate(db, { migrationsFolder: 'src/lib/db/migrations' })
+
+  const end = Date.now()
+
+  console.log("✅ Migrations completed in", end - start, "ms")
+
+  process.exit(0)
+}
+
+runMigrate().catch((err) => {
+  console.error("❌ Migration failed")
+  console.error(err)
+  process.exit(1)
+})
